test-requirements.txt
mock
pytest
pytest-cov
requests-mock
frozenlist2

mypy; python_version > "3.0"
black; python_version > "3.0"
pylint; python_version > "3.0"
coveralls; python_version > "3.0"
